Initial implementation of Checks API for Jenkins
This plugin implements the checks API provided by the Gerrit UI to
show results of CI systems on the change screen for Jenkins CIs.

The current implementation only works with MultiBranch pipeline
jobs using the Gerrit SCM provided by the GerritCodeReview plugin
for Jenkins. Other job types will be added in future changes.

One or more jenkins servers with one or more jobs can be configured
for a project. The plugin will regularly fetch jobs created for changes
by the Multibranch pipeline and will extract the build for the
currently viewed patchset.

The plugin is not able to use all features provided by the Checks
API, e.g. providing a result summary, since Jenkins does not
provide this information out of the box. This could however be enabled
in the future by extending Jnekins with a plugin.

Note, that the plugin will do CORS requests to Jenkins, since it is
running in the client's browser. Thus, CORS has to be enabled in
Jenkins.